Better handle scale factor and zoom in the tile selection window
The TileSelectionWindow now uses its own value for scaling the tilemap displayed within the window. It is completely independent of the game. In the future, it would be nice to have zoom in/out buttons at the right upper corner of the tile selection window, so the user can select how the tilemap should be displayed.